+++ src/binder_v2/create_table.rs
@@ -1,8 +1,43 @@
 use std::collections::HashSet;
+use std::result::Result as RawResult;
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
 
 use super::*;
 use crate::catalog::{ColumnCatalog, ColumnDesc};
-use crate::types::ColumnId;
+use crate::types::{ColumnId, DatabaseId, SchemaId};
+
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Serialize, Deserialize)]
+pub struct BoundTable {
+    pub database_id: DatabaseId,
+    pub schema_id: SchemaId,
+    pub table_name: String,
+    pub columns_desc: Vec<ColumnDesc>,
+    pub ordered_pk_ids: Vec<ColumnId>,
+}
+
+impl std::fmt::Display for BoundTable {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "databaseId: {}, schemaId: {}, tableName: {}, columnDesc: {:?}, orderedIds: {:?}",
+            self.database_id,
+            self.schema_id,
+            self.table_name,
+            self.columns_desc,
+            self.ordered_pk_ids
+        )
+    }
+}
+
+impl FromStr for BoundTable {
+    type Err = ();
+
+    fn from_str(_s: &str) -> RawResult<Self, Self::Err> {
+        Err(())
+    }
+}
 
 impl Binder {
     pub(super) fn bind_create_table(
@@ -73,8 +108,7 @@ impl Binder {
         let table = self.egraph.add(Node::BoundTable(BoundTable {
             database_id: db.id(),
             schema_id: schema.id(),
-            table_id: 0,
-            table_name: Some(table_name.into()),
+            table_name: table_name.into(),
             columns_desc: columns
                 .iter()
                 .map(|col| col.desc())
