+++ src/binder_v2/delete.rs
@@ -6,11 +6,8 @@ impl Binder {
         table_name: TableFactor,
         selection: Option<Expr>,
     ) -> Result {
-        self.push_context();
         let table = self.bind_table(table_name)?;
         let cond = self.bind_condition(selection)?;
-        let ret = self.egraph.add(Node::Delete([table, cond]));
-        self.pop_context();
-        Ok(ret)
+        Ok(self.egraph.add(Node::Delete([table, cond])))
     }
 }
