+++ src/binder_v2/table.rs
@@ -1,42 +1,9 @@
 // Copyright 2022 RisingLight Project Authors. Licensed under Apache-2.0.
 
-use std::result::Result as RawResult;
-use std::str::FromStr;
 use std::vec::Vec;
 
-use serde::{Deserialize, Serialize};
-
 use super::*;
-use crate::catalog::{ColumnDesc, ColumnRefId};
-use crate::types::{ColumnId, DatabaseId, SchemaId, TableId};
-
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Serialize, Deserialize)]
-pub struct BoundTable {
-    pub database_id: DatabaseId,
-    pub schema_id: SchemaId,
-    pub table_id: TableId,
-    pub table_name: Option<String>,
-    pub columns_desc: Vec<ColumnDesc>,
-    pub ordered_pk_ids: Vec<ColumnId>,
-}
-
-impl std::fmt::Display for BoundTable {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "databaseId: {}, schemaId: {}, tableId: {}, columnDesc: {:?}, orderedIds: {:?}",
-            self.database_id, self.schema_id, self.table_id, self.columns_desc, self.ordered_pk_ids
-        )
-    }
-}
-
-impl FromStr for BoundTable {
-    type Err = ();
-
-    fn from_str(_s: &str) -> RawResult<Self, Self::Err> {
-        Err(())
-    }
-}
+use crate::catalog::ColumnRefId;
 
 impl Binder {
     pub(super) fn bind_from(&mut self, tables: Vec<TableWithJoins>) -> Result {
